added `TestCheck` and moved some tests from `TestCppcheck`
